added single player fight manager

+++ Assets/Game/Scripts/Healthbar.cs
@@ -12,13 +12,13 @@ public class Healthbar : MonoBehaviour
         Slider = GetComponent<Slider>();
     }
 
-    public void Initialize(int value)
+    public void Initialize(float value)
     {
         Slider.maxValue = value;
         Slider.value = value;
     }
 
-    public void SetValue(int value)
+    public void SetValue(float value)
     {
         Slider.value = value;
     }
